feat: implement spell checker with 4 data structures and benchmarking
- Added SpellChecker interface and 4 implementations:
  - NaiveListChecker (linear list)
  - TreeSetChecker (BBST)
  - HashSetChecker (hash table)
  - TrieChecker (trie)
- Added SpellUtils.java for loading and tokenizing text/dictionary
- Added SpellBenchmark.java to measure build and check times
- Generated demo input files: english.txt (dictionary), spell_text.txt (text)
- Benchmarks output to spell_build.csv and spell_check.csv

+++ src/main/java/Main.java
@@ -1,4 +1,4 @@
-import algorithms.*;
+import patternmatch.*;
 import utils.*;
 
 import java.io.IOException;
@@ -117,6 +117,6 @@ public class Main {
             long end = System.nanoTime();
             total += (end - start);
         }
-        return total / REPEATS / 1_000_000.0;
+        return (double) total / REPEATS / 1_000_000.0;
     }
 }
